server/requirements.txt
Flask==1.1.2
Pillow==7.0.0
opencv-python==4.5.1.48
